+++ judge-impl/src/test/java/com/dwarfeng/judge/impl/service/DriverSupportMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -30,10 +31,7 @@ public class DriverSupportMaintainServiceImplTest {
     public void setUp() {
         for (int i = 0; i < 5; i++) {
             DriverSupport driverSupport = new DriverSupport(
-                    new StringIdKey("driver-support-" + (i + 1)),
-                    "label-" + (i + 1),
-                    "这是测试用的DriverSupport",
-                    "1233211234567"
+                    new StringIdKey("driver-support-" + (i + 1)), "label", "description", "exampleParam"
             );
             driverSupports.add(driverSupport);
         }
@@ -45,16 +43,20 @@ public class DriverSupportMaintainServiceImplTest {
     }
 
     @Test
-    public void test() throws Exception {
+    public void testForCrud() throws Exception {
         try {
             for (DriverSupport driverSupport : driverSupports) {
-                driverSupport.setKey(service.insert(driverSupport));
-                service.update(driverSupport);
+                driverSupport.setKey(service.insertOrUpdate(driverSupport));
                 DriverSupport testDriverSupport = service.get(driverSupport.getKey());
+                assertEquals(BeanUtils.describe(driverSupport), BeanUtils.describe(testDriverSupport));
+                testDriverSupport = service.get(driverSupport.getKey());
                 assertEquals(BeanUtils.describe(driverSupport), BeanUtils.describe(testDriverSupport));
             }
         } finally {
             for (DriverSupport driverSupport : driverSupports) {
+                if (Objects.isNull(driverSupport.getKey())) {
+                    continue;
+                }
                 service.deleteIfExists(driverSupport.getKey());
             }
         }
